BadMIM/ReconstructionHijacking/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
anyio=4.6.2=py39h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
arrow=1.3.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py39h06a4308_0
attrs=24.2.0=py39h06a4308_0
babel=2.11.0=py39h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.3=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.7=py39ha9d4c09_0
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=py39h06a4308_0
cffi=1.17.1=py39h1fdaa30_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.1=py39h06a4308_0
contourpy=1.3.0=pypi_0
crcmod=1.7=pypi_0
cryptography=43.0.3=pypi_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.0=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.3=h6a678d5_0
filelock=3.14.0=pypi_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.54.1=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.10.0=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
grad-cam=1.5.4=pypi_0
grpcio=1.67.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h11=0.14.0=py39h06a4308_0
h5py=3.12.1=pypi_0
httpcore=1.0.2=py39h06a4308_0
httpx=0.27.0=py39h06a4308_0
icu=73.1=h6a678d5_0
idna=3.10=pypi_0
imageio=2.33.1=py39h06a4308_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
importlib_metadata=7.0.1=hd3eb1b0_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=py39h06a4308_0
ipympl=0.9.4=pypi_0
ipython=8.15.0=py39h06a4308_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=py39h06a4308_0
jinja2=3.1.4=py39h06a4308_1
jmespath=0.10.0=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
json5=0.9.25=py39h06a4308_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=py39h06a4308_0
jsonschema-specifications=2023.7.1=py39h06a4308_0
jupyter-lsp=2.2.0=py39h06a4308_0
jupyter_client=8.6.0=py39h06a4308_0
jupyter_core=5.7.2=py39h06a4308_0
jupyter_events=0.10.0=py39h06a4308_0
jupyter_server=2.14.1=py39h06a4308_0
jupyter_server_terminals=0.4.4=py39h06a4308_1
jupyterlab=4.3.0=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.27.3=py39h06a4308_0
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h143b758_1
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libllvm14=14.0.6=hecde1de_4
libpng=1.6.39=h5eee18b_0
libpq=12.20=hdbd6064_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mat4py=0.6.0=pypi_0
mat73=0.65=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.6=py39h06a4308_0
mdurl=0.1.2=pypi_0
mistune=2.0.4=py39h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.10=py39h5eee18b_0
mkl_random=1.2.7=py39h1128e8f_0
mmcv=2.1.0=pypi_0
mmdet=3.3.0=pypi_0
mmengine=0.10.5=pypi_0
mmpretrain=1.2.0=pypi_0
model-index=0.1.11=pypi_0
modelindex=0.0.2=pypi_0
mpmath=1.3.0=pypi_0
mysql=5.7.24=h721c034_2
nbclient=0.8.0=py39h06a4308_0
nbconvert=7.16.4=py39h06a4308_0
nbformat=5.10.4=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=py39h06a4308_0
networkx=3.2.1=pypi_0
nodejs=20.17.0=hb8e3597_0
notebook=7.2.2=py39h06a4308_1
notebook-shim=0.2.3=py39h06a4308_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
opencv-python=4.10.0.84=pypi_0
opendatalab=0.0.10=pypi_0
openjpeg=2.5.2=he7f1fd0_0
openmim=0.3.9=pypi_0
openssl=3.0.15=h5eee18b_0
openxlab=0.1.2=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
overrides=7.4.0=py39h06a4308_0
packaging=24.1=py39h06a4308_0
pandas=2.2.2=py39h6a678d5_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.42=hebb0a14_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=11.0.0=pypi_0
pip=24.2=py39h06a4308_0
platformdirs=4.3.6=pypi_0
ply=3.11=py39h06a4308_0
prometheus_client=0.14.1=py39h06a4308_0
prompt-toolkit=3.0.43=py39h06a4308_0
protobuf=5.28.3=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.15=pypi_0
pyqt5-sip=12.15.0=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.20=he870216_1
python-dateutil=2.9.0post0=py39h06a4308_2
python-fastjsonschema=2.16.2=py39h06a4308_0
python-json-logger=2.0.7=py39h06a4308_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.4=pypi_0
pyyaml=6.0.2=py39h5eee18b_0
pyzmq=25.1.2=py39h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
referencing=0.30.2=py39h06a4308_0
requests=2.32.3=py39h06a4308_1
rfc3339-validator=0.1.4=py39h06a4308_0
rfc3986-validator=0.1.1=py39h06a4308_0
rich=13.4.2=pypi_0
rpds-py=0.10.6=py39hb02cf49_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
send2trash=1.8.2=py39h06a4308_0
setuptools=60.2.0=pypi_0
shapely=2.0.6=pypi_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py39h06a4308_0
soupsieve=2.5=py39h06a4308_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
terminado=0.17.1=py39h06a4308_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.30=pypi_0
timm=0.4.12=pypi_0
tinycss2=1.2.1=py39h06a4308_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py39h06a4308_0
torch=2.5.0=pypi_0
torchaudio=2.5.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.20.0=pypi_0
tornado=6.4.1=py39h5eee18b_0
tqdm=4.65.2=pypi_0
traitlets=5.14.3=py39h06a4308_0
triton=3.1.0=pypi_0
ttach=0.0.3=pypi_0
types-python-dateutil=2.9.0.20241003=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024b=h04d1e81_0
uri-template=1.3.0=pypi_0
urllib3=1.26.20=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=1.8.0=py39h06a4308_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yapf=0.40.2=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
